Don't work

+++ src/main/java/com/example/adressbook/repository/AdressBookRepository.java
@@ -5,8 +5,9 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AdressBookRepository extends CrudRepository<Adress, Integer>
+public interface AdressBookRepository extends CrudRepository<Adress, Long>
 {
-//    public Adress findAdressByUser_Firstname(String firstname);
+ //   public Adress findAdressByUser_FirstnameAndUser_Lastname(String firstname, String lastname);
+
 
 }
